Added Python and shell scripts to automate downloading source packages.

+++ setup.sh
@@ -19,29 +19,24 @@ if ! grep -q "$LFS" /proc/mounts; then
 	mount "$(echo $LFS_DISK)p3" "$LFS"
 	/sbin/swapon "$(echo $LFS_DISK)p2"
 fi
+if ! [ -e $(echo $LFS_DISK)p3 ]; then
+    echo "$LFS_DISK failed to mount."
+    exit 1
+fi
 
 # Download package sources and patches
 mkdir -pv $LFS/sources
 chmod -v a+wt $LFS/sources
-if ! [ -e ./wget-list ]; then
-	wget https://www.linuxfromscratch.org/lfs/view/stable/wget-list
-	sed -e "s/https:\/\/zlib\.net\/zlib-1\.2\.11\.tar\.xz/https:\/\/zlib.net\/zlib-1.2.12.tar.xz/" -i wget-list
-	wget --input-file=wget-list --continue --directory-prefix=$LFS/sources
-	wget https://www.linuxfromscratch.org/lfs/view/stable/md5sums --directory-prefix=$LFS/sources
-	sed -e "s/^.\+zlib-1\.2\.11\.tar\.xz//" -i $LFS/sources/md5sums
-	pushd $LFS/sources
-	    #md5sum -c md5sums
-	popd
-fi
+source download-packages.sh $LFS/sources
 
 # Create directories for build
 mkdir -pv $LFS/{etc,var} $LFS/usr/{bin,lib,sbin}
 for i in bin lib sbin; do
 	ln -svf usr/$i $LFS/$i
 done
-if echo $LFS_TGT | grep -qs '64'; then
-	mkdir -pv $LFS/lib64
-fi
+case $(uname -m) in
+    x86_64) mkdir -pv $LFS/lib64 ;;
+esac
 mkdir -pv $LFS/tools
 
 # Add the LFS user if it doesn't exist and give ownership of files
@@ -53,8 +48,8 @@ $LFS_USER_PW
 $LFS_USER_PW
 EOF
 chown -v lfs $LFS/{usr{,/*},lib,var,etc,bin,sbin,tools,sources}
-if echo $LFS_TGT | grep -qs '64'; then 
-	chown -v lfs $LFS/lib64 
+case $(uname -m) in
+    x86_64) chown -v lfs $LFS/lib64 ;;
+esac
 fi
-fi
-su - lfs
+#su - lfs
